@@ -1,6 +1,6 @@
 {
   "name": "nuudel-core",
-  "version": "0.1.185",
+  "version": "0.2.0",
   "description": "Web components for react with material UI",
   "main": "build/main/index.js",
   "typings": "build/main/index.d.ts",
@@ -64,7 +64,6 @@
     "graphql": "^15.7.2",
     "graphql-2-json-schema": "0.2.0",
     "i18next": "^21.6.6",
-    "material-ui-chip-input": "^1.1.0",
     "next-with-apollo": "^5.2.1",
     "nuudel-utils": "^0.1.68",
     "pica": "^9.0.1",
@@ -81,9 +80,9 @@
   "devDependencies": {
     "@ava/typescript": "^1.1.1",
     "@istanbuljs/nyc-config-typescript": "^1.0.1",
-    "@material-ui/core": "^4.12.4",
-    "@material-ui/icons": "^4.11.3",
-    "@material-ui/lab": "^4.0.0-alpha.61",
+    "@mui/icons-material": "^5.11.16",
+    "@mui/material": "^5.12.3",
+    "tss-react": "^4.8.3",
     "@react-native-community/eslint-config": "^3.0.1",
     "@types/i18next": "^13.0.0",
     "@types/node-sass": "^4.11.2",
@@ -103,7 +102,7 @@
     "eslint-plugin-import": "^2.22.0",
     "gh-pages": "^3.1.0",
     "graphql-tag": "^2.12.6",
-    "next": "^11.1.4",
+    "next": "^12.3.4",
     "node-sass": "^7.0.1",
     "npm-run-all": "^4.1.5",
     "nyc": "^15.1.0",
@@ -120,11 +119,11 @@
     "typescript": "^4.0.2"
   },
   "peerDependencies": {
-    "@material-ui/core": "^4.12.3",
-    "@material-ui/icons": "^4.11.2",
-    "@material-ui/lab": "^4.0.0-alpha.60",
+    "@mui/icons-material": "^5.11.16",
+    "@mui/material": "^5.12.3",
+    "tss-react": "^4.8.3",
     "graphql-tag": "^2.12.6",
-    "next": "^11.1.4",
+    "next": "^12.3.4",
     "react": "*",
     "swr": "^1.0.1"
   },
