@@ -1,8 +1,8 @@
 import React from 'react';
-import { Box as BaseBox, BoxProps } from '@material-ui/core';
+import { Box as BaseBox, BoxProps } from '@mui/material';
 
 interface IBoxProps extends BoxProps {}
 
-export const Box: React.FC<IBoxProps> = props => <BaseBox {...props} />;
+export const Box: React.FC<IBoxProps> = (props) => <BaseBox {...props} />;
 
 export default Box;
