@@ -1,9 +1,9 @@
 import React from 'react';
-import { Checkbox as BaseCheckbox, CheckboxProps } from '@material-ui/core';
+import { Checkbox as BaseCheckbox, CheckboxProps } from '@mui/material';
 
 interface ICheckboxProps extends CheckboxProps {}
 
-export const Checkbox: React.FC<ICheckboxProps> = props => (
+export const Checkbox: React.FC<ICheckboxProps> = (props) => (
   <BaseCheckbox {...props} />
 );
 
