@@ -1,9 +1,9 @@
 import React from 'react';
-import { Divider as BaseDivider, DividerProps } from '@material-ui/core';
+import { Divider as BaseDivider, DividerProps } from '@mui/material';
 
 interface IDividerProps extends DividerProps {}
 
-export const Divider: React.FC<IDividerProps> = props => (
+export const Divider: React.FC<IDividerProps> = (props) => (
   <BaseDivider {...props} />
 );
 
