@@ -1,21 +1,21 @@
 import React, { useEffect, useRef } from 'react';
-import {
-  makeStyles,
-  SwipeableDrawer,
-  SwipeableDrawerProps,
-} from '@material-ui/core';
+import { SwipeableDrawer, SwipeableDrawerProps } from '@mui/material';
 import clsx from 'clsx';
 import { isIOS } from 'react-device-detect';
+import { Theme } from '@mui/material/styles';
+import { makeStyles } from 'tss-react/mui';
 
-const useStyles = makeStyles({
-  list: {
-    width: 300,
-    overflowX: 'hidden',
-    height: '100%',
-  },
-  fullList: {
-    width: 'auto',
-  },
+const useStyles = makeStyles()((theme: Theme) => {
+  return {
+    list: {
+      width: 300,
+      overflowX: 'hidden',
+      height: '100%',
+    },
+    fullList: {
+      width: 'auto',
+    },
+  };
 });
 
 type Anchor = 'top' | 'left' | 'bottom' | 'right';
@@ -27,27 +27,26 @@ interface IDrawerProps extends SwipeableDrawerProps {
 
 export const Drawer: React.FC<IDrawerProps> = ({ children, ...props }) => {
   const { anchor = 'right' } = props;
-  const classes = useStyles();
+  const { classes } = useStyles();
   const [open, setOpen] = React.useState(props.open === true);
 
-  const toggleDrawer = (extend: boolean) => (
-    event: React.KeyboardEvent | React.MouseEvent,
-  ) => {
-    if (
-      event &&
-      event.type === 'keydown' &&
-      ((event as React.KeyboardEvent).key === 'Tab' ||
-        (event as React.KeyboardEvent).key === 'Shift')
-    ) {
-      return;
-    }
-    setOpen(extend);
-    if (extend && props.onOpen) {
-      props.onOpen(event);
-    } else if (!extend && props.onClose) {
-      props.onClose(event);
-    }
-  };
+  const toggleDrawer =
+    (extend: boolean) => (event: React.KeyboardEvent | React.MouseEvent) => {
+      if (
+        event &&
+        event.type === 'keydown' &&
+        ((event as React.KeyboardEvent).key === 'Tab' ||
+          (event as React.KeyboardEvent).key === 'Shift')
+      ) {
+        return;
+      }
+      setOpen(extend);
+      if (extend && props.onOpen) {
+        props.onOpen(event);
+      } else if (!extend && props.onClose) {
+        props.onClose(event);
+      }
+    };
 
   const didMountRef = useRef(false);
   useEffect(() => {
