@@ -1,9 +1,9 @@
 import React from 'react';
-import { Slider as BaseSlider, SliderProps } from '@material-ui/core';
+import { Slider as BaseSlider, SliderProps } from '@mui/material';
 
 interface ISliderProps extends SliderProps {}
 
-export const Slider: React.FC<ISliderProps> = props => (
+export const Slider: React.FC<ISliderProps> = (props) => (
   <BaseSlider {...props} />
 );
 
